==> 1_Twitter-API-Wrangling-the-Data.ipynb <==
# WRANGLING DATA WITH TWITTER API AND TWEEPY
### COMPARE THE TWEETS FROM MY GERMAN FRIENDS/FOLLOWERS WITH MY AMERICAN FRIENDS/FOLLOWERS

# Step 0:
### Importing libraries and authorizing Twitter API:

# import sys
# !{sys.executable} -m pip install tweepy
import tweepy
import json
import tweepy
import pandas as pd

# This doesnt work (taken from class notes to refresh memory)

# cred = {}
# cred['API_KEY'] = API_KEY
# cred['API_SECRET'] = API_SECRET
# cred['ACCESS_TOKEN'] = ACCESS_TOKEN
# cred['ACCESS_TOKEN_SECRET'] = ACCESS_TOKEN_SECRET

# with open('twitter_credentials copy2.json', 'w') as outfile:
#     json.dump(cred, outfile)

#note: in case this doesnt work, I changed the name of the file (before it was "twitter_credentials copy.json") to add to gitignore
with open('twitter_credentials_copy.json', 'r') as infile:
    credentials = json.loads(infile.read())

#### Twitter API - documentation:
https://docs.tweepy.org/en/latest/api.html 

help(tweepy)

# need to create an authentication object to authenticate ourselves!

auth = tweepy.OAuthHandler(credentials['API_KEY'], credentials['API_SECRET'])

# step 1: authenticate yourself
# this can be found in the authentication tutorial on the twitter developer page docs.

# step 2: set your tokens
# we need to now SET our access tokens
# this sets the tokens for you to actually be able to receive information from the API

auth.set_access_token(credentials['ACCESS_TOKEN'], credentials['ACCESS_TOKEN_SECRET'])

# step 3: generate your API
# now create our access to the API

api = tweepy.API(auth)

api.me()

# when working with complex APIs that require authentication, it is good to run a simple command like this at the very beginning
# so you can be sure that the authentication worked! 

# Step 1:
### Gather Friends and Followers to later gather their tweets:

#### Filter through my Twitter followers by their location:
https://underthecurve.github.io/jekyll/update/2017/08/02/twitter-followers.html

# api.followers?

# users = [user for user in api.followers('id', 'screen_name', count=250)]
# users

#### Make a dataframe with my Friends:

# function from SNV twitter network:

def get_friends(user):
    """ returns a list of friends (twitter_handle) of a user"""
    friends = []
    n=0
    for page in tweepy.Cursor(api.friends, screen_name=user, count=200).pages():
        n+=1
        print(f"Page number {n}")
        friends.extend(page)
#       time.sleep(60)
        return (friends)

kate_friends = get_friends("kate_saslow")

kate_friends[0]._json

names_friends = [kate_friends[i]._json['name'] for i in range(len(kate_friends))]

handles_friends = [kate_friends[i]._json['screen_name'] for i in range(len(kate_friends))]

locations_friends = [kate_friends[i]._json['location'] for i in range(len(kate_friends))]

descriptions_friends = [kate_friends[i]._json['description'] for i in range(len(kate_friends))]

friends_df = pd.DataFrame(zip(names_friends, handles_friends, locations_friends, descriptions_friends))
friends_df.rename(columns={0:'Friend', 1:'Handle', 2:'Location', 3:'UserDescription'}, inplace=True)
friends_df

# save friends_df to csv for further analysis:

friends_df.to_csv('twitter_friends_info_df.csv')

len(friends_df['Handle'])

# why are there only 20 friends? 

#### Make dataframe with my Followers:

# function from twitter SNV analysis:

def get_follower(user):
    '''Returns a list of all follower (twitter_handle) of a user
    taking into account twitter API's request frequency limits'''
    follower = []
    n=0
    for page in tweepy.Cursor(api.followers, screen_name=user, count=200).pages():
        n+=1
        print(f"Page number {n}")
        follower.extend(page)
#       time.sleep(60)
    #screen_names = [user.screen_name for user in friends]
    return (follower)

kate = get_follower('kate_saslow')
kate[0]

type(kate[0]._json)

kate[0]._json

names_followers = [kate[i]._json['name'] for i in range(len(kate))]

handles_followers = [kate[i]._json['screen_name'] for i in range(len(kate))]

locations_followers = [kate[i]._json['location'] for i in range(len(kate))]

descriptions_followers = [kate[i]._json['description'] for i in range(len(kate))]

followers_df = pd.DataFrame(zip(names_followers,handles_followers, locations_followers, descriptions_followers))
followers_df.rename(columns={0:'Follower', 1:'Handle', 2:'Location', 3:'UserDescription'}, inplace=True)
followers_df

# will eventually need to clean to standardize how the location is formatted (e.g. by country only)

# Save follower_df to csv:
followers_df.to_csv('twitter_followers_info_df.csv')

# Step 2:
### Gathering and Storing Tweets:

#### Write function that gets tweets by looping through the friends/followers in the list comprehensions above:

def get_tweets(handle):

    # extract 200 tweets
    number_tweets = 200
    tweets = api.user_timeline(screen_name=handle)

    # Empty array
    tmp = []

    # create array of tweet info: username, tweet id, date/time, text
    tweets_for_csv = [tweet.text for tweet in tweets] 

    for j in tweets_for_csv:
        tmp.append(j)
    

    return tmp

# test it to see if it worked, then apply loop to list comp.
get_tweets('SirJacobSaslow')

#### It seems that extracting the tweets' metadata worked. 
Now try it on all Followers rather than just the first 5!

handles = list(followers_df['Handle'])
handles

handles = list(followers_df['Handle'])
tweets = pd.DataFrame()

for handle in handles[0:5]:
    tweets_handle = pd.DataFrame(get_tweets(handle))
    tweets_handle['Handle'] = handle
    tweets = pd.concat([tweets, tweets_handle], axis=0)
    
tweets.head()

tweets['Handle'].value_counts()

tweets.rename(columns={0: 'Tweets'}, inplace=True)
tweets.head()        

#### What other sort of meta data can I extract from the tweets?
Consider getting date/time stamp and location of each tweet and adding it to the df

#### Extracting other meta data from tweepy internal methods:

Doesn't work on tweets df created, testing below on my own tweets:

def get_user_tweets(api, username):
        """Return a list of all tweets from the authenticated API"""
        tweets = []
        for status in tweepy.Cursor(api.user_timeline, screen_name=username).items():
            tweets.append(status)
        return tweets 

alltweets = get_user_tweets(api, username='kate_saslow')

print("Number of tweets extracted: {}.\n".format(len(alltweets)))

testtweets = get_user_tweets(api, username='SirJacobSaslow')

print("Number of tweets extracted: {}.\n".format(len(testtweets)))

print("5 recent tweets:\n")
for tweet in alltweets[:5]:
    print(tweet.text)
    print()

# We create a pandas DataFrame as follows
# Note: We loop through each element and add it to the DataFrame
data = pd.DataFrame(data=[tweet.text for tweet in alltweets], columns=['Tweets'])

# We display the first 10 elements of the DataFrame:
display(data.head(10))

print(dir(alltweets[0]))

# Print a selection of attributes from the first Tweet
print(alltweets[0].id)
print(alltweets[0].created_at)
print(alltweets[0].source)
print(alltweets[0].favorite_count)
print(alltweets[0].retweet_count)
print(alltweets[0].geo)
print(alltweets[0].coordinates)
print(alltweets[0].entities)

print(testtweets[0].id)
print(testtweets[0].created_at)
print(testtweets[0].source)
print(testtweets[0].favorite_count)
print(testtweets[0].retweet_count)
print(testtweets[0].geo)
print(testtweets[0].coordinates)
print(testtweets[0].entities)

import numpy as np
# Add attributes of interest
data['len']  = np.array([len(tweet.text) for tweet in alltweets])
data['ID']   = np.array([tweet.id for tweet in alltweets])
data['Date'] = np.array([tweet.created_at for tweet in alltweets])
data['geo'] = np.array([tweet.geo for tweet in alltweets])
data['Likes']  = np.array([tweet.favorite_count for tweet in alltweets])
data['RTs']    = np.array([tweet.retweet_count for tweet in alltweets])

# Display of first 10 elements from DataFrame
display(data.head(10))

testdf = pd.DataFrame(data=[tweet.text for tweet in testtweets], columns=['Tweets'])

testdf['len']  = np.array([len(tweet.text) for tweet in testtweets])
testdf['ID']   = np.array([tweet.id for tweet in testtweets])
testdf['Date'] = np.array([tweet.created_at for tweet in testtweets])
testdf['geo'] = np.array([tweet.geo for tweet in testtweets])
testdf['Likes']  = np.array([tweet.favorite_count for tweet in testtweets])
testdf['RTs']    = np.array([tweet.retweet_count for tweet in testtweets])

# Display of first 10 elements from DataFrame
display(testdf.head(10))

Now: before applying the for loop to the entire list of handles in the friends_df, consider what other sort of meta data I want from the tweets. Do I also want the date/time stamp? GEOCODE? etc. This way, I am only making one call to the API and getting all the meta data in one fell swoop.

# https://github.com/RodolfoFerro/pandas_twitter/blob/master/01-extracting-data.md

#### Okay, got it to work by entering a specific username. How can I have it iterate through my handles list?

# write new function that gets more data, not just .text
# see internal methods of tweepy for other meta data: https://github.com/RodolfoFerro/pandas_twitter/blob/master/01-extracting-data.md
# same link as used below!

def get_tweet_meta_data(handle):
    """ This function gets not only tweets.text but also other useful meta data using tweepy's internal method.
    Not sure I will need the other data, but this function will generate the important data in lists."""
    
    # extract 200 tweets
    number_tweets = 200
    tweets = api.user_timeline(screen_name=handle)
    
    # Empty array
    texts = []
    ids = []
    timestamps = []
    geocodes = []
    
    # create array of tweet info: username, tweet id, date/time, text, location
    tweets_text_for_csv = [tweet.text for tweet in tweets]
    tweets_id_for_csv = [tweet.id for tweet in tweets]
    tweets_createdat_for_csv = [tweet.created_at for tweet in tweets]
    tweets_geo_for_csv = [tweet.geo for tweet in tweets]
    
    for i in tweets_text_for_csv:
        texts.append(i)
        
    for j in tweets_id_for_csv:
        ids.append(j)
        
    for k in tweets_createdat_for_csv:
        timestamps.append(k)
        
    for l in tweets_geo_for_csv:
        geocodes.append(l)
        
    return texts, ids, timestamps, geocodes

get_tweet_meta_data('SirJacobSaslow')

# Now try and add to the function to add the lists into a df:

def test_get_tweet_meta_data_to_df(handle):
    """ This function (TEST function for only 5 handles!) gets not only tweets.text but also other useful meta data using tweepy's internal method.
    Not sure I will need the other data, but this function will add it to a DataFrame in one step."""
# extract 200 tweets
    number_tweets = 200
    tweets = api.user_timeline(screen_name=handle)
    tweets_meta_df = pd.DataFrame()
    
    
    # Empty array
    texts = []
    ids = []
    timestamps = []
    geocodes = []
    
    # create array of tweet info: username, tweet id, date/time, text, location
    tweets_text_for_csv = [tweet.text for tweet in tweets]
    tweets_id_for_csv = [tweet.id for tweet in tweets]
    tweets_createdat_for_csv = [tweet.created_at for tweet in tweets]
    tweets_geo_for_csv = [tweet.geo for tweet in tweets]
    
    for i in tweets_text_for_csv:
        texts.append(i)
        
    for j in tweets_id_for_csv:
        ids.append(j)
        
    for k in tweets_createdat_for_csv:
        timestamps.append(k)
        
    for l in tweets_geo_for_csv:
        geocodes.append(l)
    
    for handle in handles[0:5]:
        twts_df = pd.DataFrame(get_tweets(handle))
        twts_df['Handle'] = handle
        twts_df['ID'] = pd.Series(ids)
        twts_df['CreatedAt'] = pd.Series(timestamps)
        twts_df['Location'] = pd.Series(geocodes)
        tweets_meta_df = pd.concat([tweets_meta_df, twts_df], axis=0, ignore_index=True)
    
    return tweets_meta_df

test_tweets = test_get_tweet_meta_data_to_df(handle)

test_tweets.shape

#### After extracting tweets for all handles in Friends:
After applying the function to all handles (from the Followers df), merge dataframes with tweets.text and followers (merge on handle), to get the location as well. 

Then, clean and standardize how location is written (by country) to use for analysis.

#### Change the `[handles]` to extract tweets for FOLLOWERS, not friends to extract more tweets:

Use the Try/Except clause in function because the old function kept returning a "Not Authorized" error

def get_tweet_meta_data_to_df(handle):
    """ This function takes the test function above and applies it to all handles within the friends_df and gets not only tweets.text but also other useful meta data using tweepy's internal method.
    Not sure I will need the other data, but this function will add it to a DataFrame in one step."""

#     # extract 200 tweets
#     number_tweets = 200
    tweets = api.user_timeline(screen_name=handle)
    tweets_meta_df = pd.DataFrame()
    
    
    # Empty array
    texts = []
    ids = []
    timestamps = []
    geocodes = []
    errors = []
    
    # create array of tweet info: username, tweet id, date/time, text, location
    tweets_text_for_csv = [tweet.text for tweet in tweets]
    tweets_id_for_csv = [tweet.id for tweet in tweets]
    tweets_createdat_for_csv = [tweet.created_at for tweet in tweets]
    tweets_geo_for_csv = [tweet.geo for tweet in tweets]
    
    for i in tweets_text_for_csv:
        texts.append(i)
        
    for j in tweets_id_for_csv:
        ids.append(j)
        
    for k in tweets_createdat_for_csv:
        timestamps.append(k)
        
    for l in tweets_geo_for_csv:
        geocodes.append(l)
    
    for handle in handles:
        try:
            twts_df = pd.DataFrame(get_tweets(handle))
            twts_df['Handle'] = handle
            twts_df['ID'] = pd.Series(ids)
            twts_df['CreatedAt'] = pd.Series(timestamps)
            twts_df['Location'] = pd.Series(geocodes)
            tweets_meta_df = pd.concat([tweets_meta_df, twts_df], axis=0, ignore_index=True)
        except:
            errors.append(handle)
            print('causes error: ', errors)
            
    return tweets_meta_df

df = get_tweet_meta_data_to_df(handle)

# Tweak the function and use a try and except clause! But make sure you are storing the tweets already inside the function, not at the very end of the function.
# This will theoretically allow you to extract the maximum amount and not just automatically break the code. 

df.shape

# Step 3:
### Save tweets as dataframe to use in later analyses:

#### Save df to csv to use for NLP at a later stage:

# NOTE: now there are 2 dataframes
# one called: 'tweets_from_tweepy_small.csv' with 3982 tweets, 
# and one 'tweets_from_tweepy_large.csv' with 10,953 tweets

# CHANGE TO LARGE:
df.to_csv('tweets_from_tweepy_large.csv')
