==> 2_Cleaning-and-Restructuring-Twitter-Data.ipynb <==
# CLEANING AND RESTRUCTURING TWITTER DATA
### TAKE TWEETS FROM TWITTER API AND CLEAN/PREPARE THEM FOR ANALYSES

# Step 0:
### Importing libraries and data:

import json
import tweepy
import pandas as pd
import numpy as np

data_large = pd.read_csv("Data/tweets_from_tweepy_large.csv")

data_large.head()

data_large.shape

# 10953 tweets extracted from 554 Twitter Followers

len(data_large['Handle'])

data_large.drop('Unnamed: 0', axis=1, inplace=True)

# Import dataframe with the Friends Information (including name, handle, location, description, etc)
followers_df = pd.read_csv('Data/twitter_followers_info_df.csv')
followers_df.head()

followers_df.shape

followers_df.drop('Unnamed: 0', axis=1, inplace=True)

#### Merge the two dataframes to have Name, Handle, Tweet, and Location in one dataframe.
Merge on `'Handle'` 

df1 = data_large
df2 = followers_df

df = pd.merge(df1, df2, how='left', on='Handle')

df.shape
df.head()

df.rename(columns={'0':'TweetText', 'Location_x':'Geocode', 'Location_y':'Location'}, inplace=True)

# df.head()

df.shape

# Step 1:
### Cleaning the location data:

#### Make sure all cities/states are coded as country! Export csv files to excel, recode by hand, re-import coded locations:

locations = df['Location'].unique()

len(locations)

# Export locations df as csv, code (0: drop, 1: USA, 2: Germany) in excel, import coded locations as csv and merge with tweets df.

excel = pd.DataFrame(locations)

excel.to_csv('Data/locations_to_code.csv')

### Recode the locations in excel and import the new csv

locations_coded = pd.read_csv('Data/locations_coded.csv')

locations_coded.drop('Unnamed: 0', axis=1, inplace=True)

locations_coded.rename(columns={'0':'Location', '1':'Location_code'}, inplace=True)

locations_coded.head()

# merge tweets df and location codes:

df = df.merge(locations_coded, on='Location', how='left')

df.head()

# drop ['Location_code']==0

df = df[df['Location_code']!=0.0]

df.shape # see how many columns you are left with after dropping locations not America/Germany

df['Location_code'].value_counts()
# 5840 followers from Germany, 1087 from America. 
# Add the small csv to the analysis as well!

# Step 2:
### Adding more tweets to the analysis:

#### Because you only have 1087 tweets from the US, add the small tweets csv to the df and include in analysis as well to get more tweets from American network.

data_small = pd.read_csv('Data/tweets_from_tweepy_small.csv')
data_small.shape

data_small.drop('Unnamed: 0', axis=1, inplace=True)

# Import dataframe with the Friends Information (including name, handle, location, description, etc)
friends_df = pd.read_csv('twitter_friends_info_df.csv')

friends_df.drop('Unnamed: 0', axis=1, inplace=True)

# Merge small df and friends:

df1 = data_small
df2 = friends_df

df_small = pd.merge(df1, df2, how='left', on='Handle')

df_small.shape
df_small.head()

df_small.rename(columns={'0':'TweetText', 'Location_x':'Geocode', 'Location_y':'Location'}, inplace=True)

locations_small = df_small['Location'].unique()

locations_small = pd.DataFrame(locations_small)
locations_small.to_csv('Data/locations_sm_to_code.csv')

# Export as csv, re-code in excel, import as new csv, merge:

locations_sm_coded = pd.read_csv('Data/locations_sm_coded.csv')
locations_sm_coded.head()

locations_sm_coded.drop('Unnamed: 0', axis=1, inplace=True)

locations_sm_coded.rename(columns={'Unnamed: 2': 'Location_code'}, inplace=True)

locations_sm_coded.rename(columns={'0': 'Location'}, inplace=True)

locations_sm_coded.head()

df_small = df_small.merge(locations_sm_coded, on='Location', how='left')

df_small.head()

# Subset for where location_code != 0
df_small = df_small[df_small['Location_code']!=0.0]

df_small.shape

df_small['Location_code'].value_counts()
# 1567 German tweets
# 1040 American tweets

#### Concat df and df_small:
Then make sure to drop duplicate rows!

print("Shape of large df: ",df.shape)
print("Shape of small df: ", df_small.shape)

tweets = pd.concat([df, df_small], ignore_index=True)

tweets.shape

# Make sure there aren't duplicated rows (for Handles in both friends and followers dataframes for example.)

tweets.drop_duplicates(keep='first', inplace=True)

tweets.shape
# got rid of ~1000 duplicates

tweets['Location_code'].value_counts()

# Running your analysis on 6000 tweets from Germany and only 2000 from the US. 

# Step 3:
### Clean, fix value types, save to csv:

tweets.dtypes

tweets['Country'] = np.where((tweets['Location_code']==2.0), "Germany", "USA") 

tweets['Country'].value_counts() 
# How are these values different than those above?? 
# Don't drop ['Location_code'] just yet. 

# Step 4:
### Add Gender Dimension to Analysis:

# https://pypi.org/project/gender-guesser/

import gender_guesser.detector as gender

d=gender.Detector()
print(d.get_gender(u'Bob'))
print(d.get_gender(u'Kate'))
print(d.get_gender(u'Johanna'))
print(d.get_gender(u'Johannes'))
print(d.get_gender(u'Alex'))
print(d.get_gender(u"Jamie", u'great_britain'))

tweets['Follower'][0].split()

tweets['Follower']

def firstname(x):
    try:
        firstname = x.split(" ")[0]
    except:
        firstname = "NaN"
    
    return firstname

firstname(tweets['Follower'][0])

tweets['First_Name'] = tweets['Follower'].apply(firstname)

tweets.head()

tweets['Gender'] = tweets['First_Name'].apply(d.get_gender)
    

tweets['Gender'].value_counts()

# export to csv for excel re-coding
# import new csv from excel:



tweets = pd.read_csv('Data/tweets_to_preprocess_w_gender.csv')

tweets['Gender'].value_counts()

# save as csv just in case: 
tweets.to_csv('Data/tweets_to_preprocess_NLP.csv')
