==> 3_Sentiment-Analysis-on-Twitter-Network.ipynb <==
# SENTIMENT ANALYSIS ON TWITTER DATA

### IS MY GERMAN OR AMERICAN TWITTER NETWORK "HAPPIER" ?

# Step 0:
### Importing libraries and data:

import pandas as pd
import nltk
# nltk.download('punkt')
# nltk.download('wordnet')
# nltk.download('stopwords')
# nltk.download('vader_lexicon') 
from nltk.sentiment.vader import SentimentIntensityAnalyzer

import re
import string
import warnings
warnings.filterwarnings("ignore")
from textblob import TextBlob
from textblob_de import TextBlobDE
import matplotlib.pyplot as plt
import seaborn as sns
# sns.set_palette('Set3')
# sns.set_style("whitegrid")

# import df with tweets from previous notebook:
tweets = pd.read_csv('Data/tweets_to_preprocess_NLP.csv')

tweets.drop('Unnamed: 0', axis=1, inplace=True)

tweets.drop('Geocode', axis=1, inplace=True)

tweets.drop(['Location_code', 'Friend', 'First_Name'], axis=1, inplace=True)

tweets.head()

# Step 1:
### Before pre-processing the data for NLP, separate tweet by language:

from langdetect import detect

detect(tweets['TweetText'][0])
# it detects the first tweet as being German

detect(tweets['TweetText'][1])
# and it detects the second tweet as being English. 

for index, row in tweets['TweetText'].iteritems():
    try:
        lang = detect(row)
        
    except:
        lang = 0
        
    tweets.loc[index, 'Language'] = lang

# tweets['Language'].value_counts()

# tweets.drop(['Friend', 'Location_code'], axis=1, inplace=True)

# tweets.head()

### Basic Text Cleanup:

Remove RT, Punctuation, etc.

# add new column in df to clean up the texts
tweets['text'] = tweets['TweetText']

tweets['text']

list_tweets = list(tweets['text'])

test= tweets.iloc[0][0]

test

remove_rt = lambda x: re.sub('RT @\w+: ', "", x)

lab = lambda x: re.sub('[^äöüÄÖÜA-Za-z ]+','', x)

test1 = remove_rt(test)
test1

test2 = lab(test1)

test2.lower()

# Make this into a function:

def clean_up(x):
    """ Clean up the text before NLP. 
    Remove Punctuation, remove retweet and @.
    Keep German characters as vowels."""
    
    remove_rt = lambda x: re.sub('RT @\w+: ', "", x)
    lab = lambda x: re.sub('[^äöüÄÖÜA-Za-z ]+','', x)
    tw_1 = remove_rt(x)
    tw_2 = lab(tw_1)
    tw_3 =  tw_2.lower()
    
    return tw_3
    

# Test the function with 1 tweet before applying to the column
clean_up('RT @z_edian: Unser @snv_berlin Expert:innenworkshop zur "staatlichen Beantwortung von böswilligen Cyberaktivitäten" geht langsam zuende.\n\nD…')

tweets['text'] = tweets['text'].apply(clean_up)

# tweets.head()

### Tokenization:

def tokenize(x):
    """Tokenize the tweets."""
    
    from nltk.tokenize import word_tokenize
    tokenized_x = word_tokenize(x)
    return tokenized_x

tokenize(tweets['text'][0])

tweets['tokenized_tw'] = tweets['text'].apply(tokenize)

tweets.shape

### Subset in German and English dataframes for the stemming/lemmatizing/vectorizing/analysis:

tweets_en = tweets[tweets['Language']=='en']
tweets_en.shape

tweets_de = tweets[tweets['Language']=='de']
tweets_de.shape

# So there are now a couple hundred tweets dropped from my analysis since the language was either not detected, or not detected as en/de.

### Stemming and Lemmatization:

def stem_lemmatize_en(l):
    """This function performs both stemming and lemmatization on our english tweets."""
    
    
    from nltk.stem import WordNetLemmatizer
    lemmatizer = WordNetLemmatizer()
    l = [lemmatizer.lemmatize(item) for item in l]
    from nltk.stem import SnowballStemmer
    stemmer = SnowballStemmer("english")
    l = [stemmer.stem(item) for item in l]
    
    return l 

tweets_en['stem_lemm_tw'] = tweets_en['tokenized_tw'].apply(stem_lemmatize_en)

# tweets_en.head()
# Does it actually look like the texts have changed enough from stemming and lemmatization?? 

def stem_lemmatize_de(l):
    """This function performs both stemming and lemmatization on our german tweets."""
    
    
    from nltk.stem import WordNetLemmatizer
    lemmatizer = WordNetLemmatizer()
    l = [lemmatizer.lemmatize(item) for item in l]
    from nltk.stem import SnowballStemmer
    stemmer = SnowballStemmer("german")
    l = [stemmer.stem(item) for item in l]
    
    return l 

tweets_de['stem_lemm_tw'] = tweets_de['tokenized_tw'].apply(stem_lemmatize_de)
# tweets_de.head()

### Stopwords Removal:

def remove_stopwords_en(l):
    """Remove english stopwords from a list of tweets."""
    
    from nltk.corpus import stopwords
    stopwords_list = stopwords.words("english")
    l = [item for item in l if not item in stopwords_list]
    
    return l

tweets_en['remove_stopwords_tw'] = tweets_en['stem_lemm_tw'].apply(remove_stopwords_en)
# tweets_en.head()

def remove_stopwords_de(l):
    """Remove german stopwords from a list of tweets."""
    
    from nltk.corpus import stopwords
    stopwords_list = stopwords.words("german")
    l = [item for item in l if not item in stopwords_list]
    
    return l

tweets_de['remove_stopwords_tw'] = tweets_de['stem_lemm_tw'].apply(remove_stopwords_de)
# tweets_de.head()

### Remove intermediary processing steps from dfs:

tweets_en.drop(['tokenized_tw', 'stem_lemm_tw', 'text', 'UserDescription', 'CreatedAt', 'ID', 'Follower'], axis=1, inplace=True)
tweets_en.rename(columns={'remove_stopwords_tw': 'text_processed'}, inplace=True)
tweets_en.head()

tweets_de.drop(['tokenized_tw', 'stem_lemm_tw', 'text', 'UserDescription', 'CreatedAt', 'ID', 'Follower'], axis=1, inplace=True)
tweets_de.rename(columns={'remove_stopwords_tw': 'text_processed'}, inplace=True)
tweets_de.head()

# Step 2: 
### Sentiment Analysis with TextBlob

Calculating Negative, Positive, and/or Neutral Values.
* REMEMBER - the tweets in tweets_en means that they are English-language tweets, NOT that they come from my American network. 
* REMEMBER - the tweets in tweets_de means that they are German-language tweets, NOT that they come from my German network.

#### Using Python package TextBlob to calculate polarity values of individual tweets:

sentiment_objects_en = [TextBlob(tweet) for tweet in tweets_en['TweetText']]

sentiment_objects_en[0].polarity, sentiment_objects_en[0]

# Add polarity column to the dataframes:

tweets_en['Polarity'] = [i.polarity for i in sentiment_objects_en]
# tweets_en.head()

sentiment_objects_de = [TextBlobDE(tweet) for tweet in tweets_de['TweetText']]

sentiment_objects_de[0].polarity, sentiment_objects_de[0]

tweets_de['Polarity'] = [i.polarity for i in sentiment_objects_de]
# tweets_de.head()

### Add polarity column to df of all tweets and plot polarity by location rather than language.

#### Using Python package TextBlob to calculate sentiment (pos/neg) values of individual tweets:

TextBlob sentiment returns polarity score and subjectivity score. 
If polarity > 0, it is considered positive, if  polarity < 0, it is considered negative.

sentiment_objects_en = [TextBlob(tweet) for tweet in tweets_en['TweetText']]

sentiment_objects_en[0].sentiment, sentiment_objects_en[0]

## Add column w. subjectivity score to each df
tweets_en['Subjectivity'] = [i.sentiment[1] for i in sentiment_objects_en]
# tweets_en.head()

tweets_de['Subjectivity'] = [i.sentiment[1] for i in sentiment_objects_de]
# tweets_de.head()

#### Add column with sentiment (pos/neg/neu)

sentiment_category_en = ['positive' if score >0
                         else 'negative' if score<0
                             else 'neutral'
                                  for score in tweets_en['Polarity']]

sentiment_category_en[:20] 
# Make sure it worked, then add to df

tweets_en['Sentiment'] = sentiment_category_en

# tweets_en.head()

sentiment_category_de = ['positive' if score >0
                         else 'negative' if score<0
                             else 'neutral'
                                  for score in tweets_de['Polarity']]

sentiment_category_de[:20] 
# Make sure it worked, then add to df

tweets_de['Sentiment'] = sentiment_category_de

tweets_de.head()

# Step 3:
### Putting the en/de tweets back into one dataframe for continued analysis:

print(len(tweets_en))
print(len(tweets_de))

# Save both en/de dataframes to csv for wordcloud in separate analysis:
tweets_en.to_csv('Data/tweets_en.csv')
tweets_de.to_csv('Data/tweets_de.csv')

tweets_analysis = pd.concat([tweets_en, tweets_de], ignore_index=True)

# save df to csv:

tweets_analysis.to_csv('Data/tweets_analysis.csv')

tweets_analysis['Country'].value_counts()

type(tweets_analysis['text_processed'][0][1])

tweets_analysis['text_processed']

#### Since absolute numbers aren't telling, maybe look at ratio of positive:total tweets.

# Get length of German-region tweets for ratio:

ger = tweets_analysis[tweets_analysis['Country']=='Germany']
print("Tweets from German Network: ", len(ger))

usa = tweets_analysis[tweets_analysis['Country']=='USA']
print("Tweets from American Network: ", len(usa))

ratio_us_pos = len(tweets_analysis[(tweets_analysis['Country']=='USA') & (tweets_analysis['Sentiment']=='positive')])/len(usa)
ratio_us_pos

ratio_de_pos = len(tweets_analysis[(tweets_analysis['Country']=='Germany') & (tweets_analysis['Sentiment']=='positive')])/len(ger)
ratio_de_pos

# Step 4:
### Understand relative sentimiment rather than absolute numbers:

# Try Groupby country:
tweets_analysis.groupby(['Country']).mean()

# Overall, while average tweets tend to be neutral, American tweets tend to be more SUBJECTIVE. 

grouped_country = tweets_analysis.groupby(['Country', 'Sentiment']).mean()
grouped_country

# Overall, American tweets tend to be more SUBJECTIVE, in both positive and negative tweets. 

grouped_gender = tweets_analysis.groupby(['Gender', 'Sentiment']).mean()
grouped_gender

# Cross Tabs 
rel_values = pd.crosstab(index=tweets_analysis['Country'], columns=tweets_analysis['Sentiment'])

for row in rel_values.iterrows():
    rel_values['percent_of_pos'] = rel_values['positive']/(rel_values['negative']+rel_values['neutral']+rel_values['positive'])

rel_values
